feat: better text types
