+++ packages/text/index.ts
@@ -1,2 +1,3 @@
+export * from './constants'
 export { default as Text } from './Text'
-export * from './types'
+export * from './Text'
